Fix broken error handling
Certain errors aren't handled properly. In such cases the the callback is called like so `callback(null, undefined)`, while the http response actually contains an object with an `__error` field, which isn't checked.
This commit addresses that issue by setting `err` to `res.__error` in case `err` evaluates to false (read: isn't set to an error yet,) and `res.__error` evaluates to `true`.